requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ML
torch>=2.2.0
torchvision>=0.17.0
xgboost>=2.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
Pillow>=10.0.0
numpy>=1.26.0
pandas>=2.2.0

# LangChain RAG
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pypdf>=4.0.0

# HTTP (Ollama client)
requests>=2.31.0


